requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.9.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-decouple==3.8
